# docs/openapi/pn-io-connector-internal.yaml
openapi: 3.0.3
info:
  title: pn-io-connector API
  version: 1.0.0
  description: API esposta da pn-io-connector per le operazioni verso App IO.

tags:
  - name: Io
    description: Operazioni verso App IO

paths:
  /io/message:
    post:
      tags: [Io]
      summary: Presa in carico di un messaggio verso App IO
      operationId: sendIOMessage
      description: >
        Endpoint sincrono invocato da client pagopa per richiedere l'invio
        di un messaggio a un cittadino tramite App IO. Il microservizio
        verifica, salvo errori di validazione, inserisce
        la richiesta nella coda di elaborazione.
      parameters:
        - $ref: '#/components/parameters/xPagopaIoConCxId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
      responses:
        '200':
          description: Richiesta presa in carico per elaborazione asincrona.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '204':
          description: Richiesta già ricevuta con stesso requestId e stesso payload (idempotenza). Nessun contenuto restituito.
        '400':
          description: Richiesta malformata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: requestId già esistente ma con payload diverso da quello originale.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /io/profile:
    post:
      tags: [Io]
      summary: Verifica raggiungibilità profilo IO del destinatario
      description: >
        Endpoint invocato dal client pagopa (es. orchestratore SEND) per
        verificare se un cittadino è raggiungibile tramite App IO (profilo presente
        e sender_allowed = true).La risposta riporta status 
        (SENDER_ALLOWED / SENDER_NOT_ALLOWED) e, se il profilo è attivo, 
        le preferredLanguages restituite da IO.
      operationId: getIOProfile
      parameters:
        - $ref: '#/components/parameters/xPagopaIoConCxId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetProfileRequest'
      responses:
        '200':
          description: Esito della verifica profilo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProfileResponse'
        '400':
          description: Richiesta malformata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /messages/{id}:
    get:
      summary: Recupero metadati allegati di un messaggio
      description: >
        Endpoint invocato da BackEnd IO quando l'utente apre il dettaglio del messaggio
        su App IO. Viene verificato che il codice fiscale del destinatario
        sia coerente con quello associato al messaggio, richiesta a SafeStorage
        una PresignedUri per ciascun allegato e restituita una lista di metadati che
        potrà essere usata dal client App IO per richiedere il contenuto binario.
      operationId: getMessage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: requestId — identificativo di correlazione remota passato da pn-io-connector in third_party_data.id al momento dell'invio del messaggio
        - name: x-pagopa-cx-taxid
          in: header
          required: true
          schema:
            type: string
          description: Codice fiscale anonimizzato (internalId) iniettato dall'API Gateway da context.authorizer.cx_id dopo la validazione Lollipop. Confrontato con il destinatario del messaggio per il controllo di coerenza.
      responses:
        '200':
          description: Metadati degli allegati associati al messaggio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMessageResponse'
        '403':
          description: Codice fiscale del chiamante non coerente con il destinatario.
        '404':
          description: Messaggio non trovato o allegati non presenti.

components:
  parameters:
    xPagopaIoConCxId:
      name: x-pagopa-iocon-cx-id
      in: header
      required: true
      description: Identificativo del client che invoca il microservizio (cxId).
      schema:
        type: string

  schemas:

    MessageRequest:
      type: object
      required:
        - requestId
        - recipientTaxId
        - senderServiceId
        - subject
        - markdown
      properties:
        requestId:
          type: string
          description: Identificativo univoco della richiesta (generato dall'orchestratore SEND).
          example: "REQ-20260408-ABC123"
        iun:
          type: string
          description: Identificativo Univoco Notifica su SEND.
          example: "ABCD-EFGH-1234-5678-X"
        recipientTaxId:
          type: string
          description: Codice fiscale anonimizzato del destinatario del messaggio.
          example: "ABCDEFG1234567890"
        senderServiceId:
          type: string
          description: Id del servizio per il quale si invia una notifica.
          example: "000000000"
        subject:
          type: string
          description: Oggetto del messaggio (max 120 caratteri).
          maxLength: 120
        markdown:
          type: string
          description: >
            Corpo del messaggio in formato Markdown. Puo contenere sezioni
            in lingue diverse concatenate nel medesimo campo.
        attachments:
          type: array
          items:
            type: string
          description: Lista dei documenti allegati al messaggio.
          nullable: true
        sensitiveContent:
          type: boolean
          description: Indica se il messaggio contiene informazioni o dati sensibili.
          example: false
          nullable: true
        dueDate:
          type: string
          format: UTCISODateFromString
          description: Permette di associare al messaggio un promemoria. Il formato data deve essere ISO-8601 e fuso orario UTC
          example: 2022-09-30T22:00:00Z
        paymentData:
          $ref: '#/components/schemas/PaymentData'
        pollingMaxHours:
          type: integer
          description: >
            Numero massimo di ore per cui effettuare il polling degli esiti.
            Determinato dall'orchestratore. Default configurato: 48h.
          default: 48

    PaymentData:
      type: object
      required:
        - amount
        - noticeCode
        - creditorTaxId
      properties:
        amount:
          type: integer
          description: Importo in centesimi di euro.
          example: 1500
        noticeCode:
          type: string
          description: Codice avviso pagoPA.
          example: "302000100440009424"
        creditorTaxId:
          type: string
          description: Codice fiscale dell'ente creditore.
          example: "01234567890"
        invalidAfterDueDate:
          type: boolean
          description: In app visualizza il pagamento come scaduto se la data attuale è successiva a dueDate
    MessageResponse:
      type: object
      properties:
        xPagopaIoConCxId:
          type: string
        requestId:
          type: string
        status:
          type: string
          enum: [ACCEPTED]
          description: Richiesta di invio presa in carico.
    GetProfileRequest:
      type: object
      required:
        - recipientTaxId
        - senderServiceId
      properties:
        recipientTaxId:
          type: string
          description: Codice fiscale anonimizzato del destinatario.
          example: "ABCDEFG1234567890"
        senderServiceId:
          type: string
          description: >
            Id del servizio IO dell'ente mittente. Necessario per recuperare la ApiKey USE.
          example: "000000000"
    GetProfileResponse:
      type: object
      properties:
        status:
          type: string
          enum: [SENDER_ALLOWED, SENDER_NOT_ALLOWED]
          description: >
            SENDER_ALLOWED = profilo presente su IO e sender_allowed = true.
            SENDER_NOT_ALLOWED = profilo assente (HTTP 404 da IO) o sender_allowed = false.
        preferredLanguages:
          type: array
          nullable: true
          description: >
            Preferenze di lingua del destinatario..
            Valorizzato solo quando status = SENDER_ALLOWED.
          items:
            type: string
            example: "it_IT"
    GetMessageResponse:
      type: object
      description: Dettagli del messaggio e lista degli allegati esposti a BackEnd IO.
      properties:
        details:
          type: object
          properties:
            subject:
              type: string
              description: Titolo del messaggio.
            markdown:
              type: string
              description: Corpo del messaggio in formato markdown.
        attachments:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Identificativo dell'allegato.
              name:
                type: string
                maxLength: 127
                description: Nome allegato con desinenza ".pdf" obbligatoria.
              content_type:
                type: string
                description: MIME type dell'allegato. Deve essere "application/pdf".
                default: application/pdf
              url:
                type: string
                format: uri
                description: URL logico dell'allegato.
              category:
                type: string
                description: Categoria dell'allegato. Deve essere "DOCUMENT".
                default: DOCUMENT
    ProblemDetail:
      type: object
      properties:
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
        traceId:
          type: string